Move promise polyfill into polyfill subdirectory. NFC
This lays the groundwork to add more polyfills over time.

Since its preferable to use standard polyfills (not authored as part of
emscripten), we exclude this directory from lint checking (i.e.  we
treat it as third-party code).

Split out from #15822
