+++ .eslintrc.yml
@@ -11,6 +11,7 @@ ignorePatterns:
   - "cache/"
   - "third_party/"
   - "tests/"
+  - "src/polyfill/"
   - "src/library*.js"
   - "src/runtime_*.js"
   - "src/shell*.js"
@@ -30,7 +31,6 @@ ignorePatterns:
   - "src/memoryprofiler.js"
   - "src/support.js"
   - "src/gl-matrix.js"
-  - "src/promise_polyfill.js"
   - "src/headless.js"
   - "src/headlessCanvas.js"
   - "src/emscripten-source-map.min.js"
