+++ src/shell.js
@@ -40,7 +40,7 @@ var Module = typeof {{{ EXPORT_NAME }}} !== 'undefined' ? {{{ EXPORT_NAME }}} : 
 // Include a Promise polyfill for legacy browsers. This is needed either for
 // wasm2js, where we polyfill the wasm API which needs Promises, or when using
 // modularize which creates a Promise for when the module is ready.
-#include "promise_polyfill.js"
+#include "polyfill/promise.js"
 #endif
 
 // See https://caniuse.com/mdn-javascript_builtins_object_assign
